requirements.txt
alembic==1.4.0
certifi==2019.11.28
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
wincertstore==0.2
WTForms==2.2.1